+++ HPXMLtoOpenStudio/tests/test_enclosure.rb
@@ -810,7 +810,7 @@ class HPXMLtoOpenStudioEnclosureTest < MiniTest::Test
     model = OpenStudio::Model::Model.new
 
     # get arguments
-    args_hash['output_dir'] = 'tests'
+    args_hash['output_dir'] = File.dirname(__FILE__)
     arguments = measure.arguments(model)
     argument_map = OpenStudio::Measure.convertOSArgumentVectorToMap(arguments)